requirements.txt
streamlit>=1.32.0
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0
numpy>=1.26.0
scipy>=1.12.0
plotly>=5.20.0
Pillow>=10.3.0
scikit-learn>=1.4.0
pandas>=2.2.0
matplotlib>=3.8.0
datasets>=2.18.0
